Refactor code into ads.llm.autogen.v02
